check sendingtransactions just once after hard refresh
